Extract message widget
