+++ components/espaper_dashboard/espaper_dashboard_widget.cpp
@@ -138,28 +138,6 @@ std::string WeatherWidget::condition_to_icon_(WeatherCondition condition) {
 
 std::string WeatherWidget::condition_to_icon_(std::string condition) {
     return this->condition_to_icon_(this->str_to_condition_(condition));
-
-
-}void MessageWidget::draw(int start_x, int start_y) {
-    display::Display *it = this->get_display_();
-
-
-    if(this->icon_.has_value()) {
-        // icon
-        it->printf(start_x+this->width_.value()/10, start_y+this->height_.value()/2, this->target_->get_large_glyph_font(), this->target_->get_light_color(), display::TextAlign::CENTER, "%s", this->icon_.value().c_str());
-        // message
-        it->printf(start_x+this->width_.value()*3/10, start_y+this->height_.value()/2, this->target_->get_large_font(), this->target_->get_foreground_color(), display::TextAlign::CENTER_LEFT, "%s", this->message_.value().c_str());
-    } else {
-        // message
-        it->printf(start_x+this->width_.value()/10, start_y+this->height_.value()/2, this->target_->get_large_font(), this->target_->get_foreground_color(), display::TextAlign::CENTER_LEFT, "%s", this->message_.value().c_str());
-    }
-
-}
-
-void MessageWidget::init_size() {
-    ESP_LOGD(TAG, "width has value: %s (%d), height has value: %s", YESNO(this->width_.has_value()), this->width_.value(), YESNO(this->height_.has_value()));
-    if(!this->width_.has_value() || !this->width_.value()) this->width_ = this->get_display_()->get_width();
-    if(!this->height_.has_value() || !this->height_.value()) this->height_ = this->width_.value()/4;
 }
 
 }
